refactor: move command contracts out of cli surface
